crm_backend/requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.0
python-dotenv==1.0.1
Werkzeug==3.1.2
email-validator==2.1.1
bcrypt==4.1.2
